requirements.txt
fastapi==0.85.1
uvicorn==0.18.3
requests==2.28.1
beautifulsoup4==4.11.1
redis==5.0.0  # or latest stable version
pydantic==1.10.2
python-dotenv==1.0.0
aiohttp==3.8.5
aiofiles==0.8.0